Remuved warnings from gzip.hpp.
Review: https://reviews.apache.org/r/53709

+++ 3rdparty/stout/include/stout/gzip.hpp
@@ -189,7 +189,7 @@ inline Try<std::string> compress(
   z_stream_s stream;
   stream.next_in =
     const_cast<Bytef*>(reinterpret_cast<const Bytef*>(decompressed.data()));
-  stream.avail_in = decompressed.length();
+  stream.avail_in = static_cast<unsigned int>(decompressed.length());
   stream.zalloc = Z_NULL;
   stream.zfree = Z_NULL;
   stream.opaque = Z_NULL;
